# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Patient-Doctor Appointment System API",
    "version": "1.0.0",
    "description": "API documentation for the Patient-Doctor Appointment System"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "type": "string", "enum": ["patient", "doctor"] }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "patientId": { "type": "string", "format": "uuid" },
          "doctorId": { "type": "string", "format": "uuid" },
          "categoryId": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": ["pending", "accepted", "completed", "cancelled"]
          },
          "appointmentDate": { "type": "string", "format": "date-time" },
          "prescription": { "type": "string" },
          "patient": { "$ref": "#/components/schemas/User" },
          "doctor": { "$ref": "#/components/schemas/User" },
          "category": { "$ref": "#/components/schemas/Category" }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Appointment" }
          },
          "total": { "type": "integer" },
          "currentPage": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      }
    }
  },
  "paths": {
...
    "/users/doctors": {
      "get": {
        "tags": ["Doctors"],
        "summary": "Get all doctors with pagination and filtering",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "in": "query",
            "name": "searchTerm",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of doctors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["success"] },
                    "data": { "$ref": "#/components/schemas/PaginatedResponse" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/doctors/available": {
      "get": {
        "tags": ["Doctors"],
        "summary": "Get available doctors for a specific date",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available doctors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["success"] },
                    "data": { "$ref": "#/components/schemas/PaginatedResponse" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
} 